Added ability to shutdown remote Windows systems
Using NET RPC you can now shutdown your Windows systems in addition to
waking up systems. It will require to submit a username and password of
a windows administrator account on the machine you'll want to shutdown
in question.
For this to work on systems newer then XP you will need to allow ICMP
connections in the Windows Firewall.